+++ R/LearnerClustAgnes.R
@@ -32,21 +32,22 @@ LearnerClustAgnes = R6Class("LearnerClustAgnes",
           tags = "train"
         ),
         trace.lev = p_int(0L, default = 0L, tags = "train"),
-        k = p_int(1L, default = 2L, tags = "predict"),
-        par.method = p_uty(tags = "train", custom_check = crate(function(x) {
-          if (!(test_numeric(x) || test_list(x))) {
-            return("`par.method` needs to be a numeric vector")
-          }
-          if (length(x) %in% c(1L, 3L, 4L)) {
-            TRUE
-          } else {
-            "`par.method` needs be of length 1, 3, or 4"
-          }
-        }))
+        k = p_int(1L, default = 2L, tags = c("train", "predict")),
+        par.method = p_uty(
+          tags = "train",
+          depends = quote(method %in% c("flexible", "gaverage")),
+          custom_check = crate(function(x) {
+            if (!(test_numeric(x) || test_list(x))) {
+              return("`par.method` needs to be a numeric vector")
+            }
+            if (length(x) %in% c(1L, 3L, 4L)) {
+              TRUE
+            } else {
+              "`par.method` needs be of length 1, 3, or 4"
+            }
+          })
+        )
       )
-
-      # param deps
-      param_set$add_dep("par.method", "method", CondAnyOf$new(c("flexible", "gaverage")))
 
       param_set$set_values(k = 2L)
 
@@ -65,16 +66,20 @@ LearnerClustAgnes = R6Class("LearnerClustAgnes",
   private = list(
     .train = function(task) {
       pv = self$param_set$get_values(tags = "train")
-      m = invoke(cluster::agnes, x = task$data(), diss = FALSE, .args = pv)
+      m = invoke(cluster::agnes,
+        x = task$data(),
+        diss = FALSE,
+        .args = remove_named(pv, "k")
+      )
       if (self$save_assignments) {
-        self$assignments = stats::cutree(m, self$param_set$values$k)
+        self$assignments = stats::cutree(m, pv$k)
       }
-
-      return(m)
+      m
     },
 
     .predict = function(task) {
-      if (self$param_set$values$k > task$nrow) {
+      pv = self$param_set$get_values(tags = "predict")
+      if (pv$k > task$nrow) {
         stopf("`k` needs to be between 1 and %i", task$nrow)
       }
 
@@ -85,5 +90,5 @@ LearnerClustAgnes = R6Class("LearnerClustAgnes",
   )
 )
 
-#' @include aaa.R
-learners[["clust.agnes"]] = LearnerClustAgnes
+#' @include zzz.R
+register_learner("clust.agnes", LearnerClustAgnes)