+++ R/LearnerClustCMeans.R
@@ -33,7 +33,7 @@ LearnerClustCMeans = R6Class("LearnerClustCMeans",
         dist = p_fct(levels = c("euclidean", "manhattan"), default = "euclidean", tags = "train"),
         method = p_fct(levels = c("cmeans", "ufcl"), default = "cmeans", tags = "train"),
         m = p_dbl(1, default = 2, tags = "train"),
-        rate.par = p_dbl(0, 1, tags = "train"),
+        rate.par = p_dbl(0, 1, tags = "train", depends = quote(method == "ufcl")),
         weights = p_uty(default = 1L, tags = "train", custom_check = crate(function(x) {
           if (test_numeric(x) && all(x > 0) || check_count(x, positive = TRUE)) {
             TRUE
@@ -43,8 +43,6 @@ LearnerClustCMeans = R6Class("LearnerClustCMeans",
         })),
         control = p_uty(tags = "train")
       )
-      # add deps
-      param_set$add_dep("rate.par", "method", CondEqual$new("ufcl"))
 
       param_set$set_values(centers = 2L)
 
@@ -62,20 +60,19 @@ LearnerClustCMeans = R6Class("LearnerClustCMeans",
   ),
   private = list(
     .train = function(task) {
-      check_centers_param(self$param_set$values$centers, task, test_data_frame, "centers")
-
       pv = self$param_set$get_values(tags = "train")
+      assert_centers_param(pv$centers, task, test_data_frame, "centers")
+
       m = invoke(e1071::cmeans, x = task$data(), .args = pv, .opts = allow_partial_matching)
       if (self$save_assignments) {
         self$assignments = m$cluster
       }
-
-      return(m)
+      m
     },
 
     .predict = function(task) {
-      partition = unclass(cl_predict(self$model, newdata = task$data(), type = "class_ids"))
-      prob = unclass(cl_predict(self$model, newdata = task$data(), type = "memberships"))
+      partition = unclass(invoke(cl_predict, self$model, newdata = task$data(), type = "class_ids"))
+      prob = unclass(invoke(cl_predict, self$model, newdata = task$data(), type = "memberships"))
       colnames(prob) = seq_len(ncol(prob))
 
       PredictionClust$new(task = task, partition = partition, prob = prob)
@@ -83,5 +80,5 @@ LearnerClustCMeans = R6Class("LearnerClustCMeans",
   )
 )
 
-#' @include aaa.R
-learners[["clust.cmeans"]] = LearnerClustCMeans
+#' @include zzz.R
+register_learner("clust.cmeans", LearnerClustCMeans)