+++ R/LearnerClustEM.R
@@ -54,21 +54,20 @@ LearnerClustEM = R6Class("LearnerClustEM",
     .train = function(task) {
       pv = self$param_set$get_values(tags = "train")
       names(pv) = chartr("_", "-", names(pv))
-      ctrl = do.call(RWeka::Weka_control, pv)
+      ctrl = invoke(RWeka::Weka_control, .args = pv)
       m = invoke(RWeka::make_Weka_clusterer("weka/clusterers/EM"), x = task$data(), control = ctrl)
       if (self$save_assignments) {
         self$assignments = unname(m$class_ids + 1L)
       }
-
-      return(m)
+      m
     },
 
     .predict = function(task) {
-      partition = predict(self$model, newdata = task$data(), type = "class") + 1L
+      partition = invoke(predict, self$model, newdata = task$data(), type = "class") + 1L
       PredictionClust$new(task = task, partition = partition)
     }
   )
 )
 
-#' @include aaa.R
-learners[["clust.em"]] = LearnerClustEM
+#' @include zzz.R
+register_learner("clust.em", LearnerClustEM)