+++ R/LearnerClustMeanShift.R
@@ -50,17 +50,16 @@ LearnerClustMeanShift = R6Class("LearnerClustMeanShift",
   ),
   private = list(
     .train = function(task) {
-      if (!is.null(self$param_set$values$subset) && length(self$param_set$values$subset) > task$nrow) {
+      pv = self$param_set$get_values(tags = "train")
+      if (!is.null(pv$subset) && length(pv$subset) > task$nrow) {
         stopf("`subset` length must be less than or equal to number of observations in task")
       }
 
-      pv = self$param_set$get_values(tags = "train")
       m = invoke(LPCM::ms, X = task$data(), .args = pv)
       if (self$save_assignments) {
         self$assignments = m$cluster.label
       }
-
-      return(m)
+      m
     },
 
     .predict = function(task) {
@@ -71,5 +70,5 @@ LearnerClustMeanShift = R6Class("LearnerClustMeanShift",
   )
 )
 
-#' @include aaa.R
-learners[["clust.meanshift"]] = LearnerClustMeanShift
+#' @include zzz.R
+register_learner("clust.meanshift", LearnerClustMeanShift)