+++ R/LearnerClustOPTICS.R
@@ -25,23 +25,23 @@ LearnerClustOPTICS = R6Class("LearnerClustOPTICS",
         eps = p_dbl(0, special_vals = list(NULL), default = NULL, tags = "train"),
         minPts = p_int(0L, default = 5L, tags = "train"),
         search = p_fct(levels = c("kdtree", "linear", "dist"), default = "kdtree", tags = "train"),
-        bucketSize = p_int(1L, default = 10L, tags = "train"),
+        bucketSize = p_int(1L, default = 10L, tags = "train", depends = quote(search == "kdtree")),
         splitRule = p_fct(
-          levels = c("STD", "MIDPT", "FAIR", "SL_MIDPT", "SL_FAIR", "SUGGEST"), default = "SUGGEST", tags = "train"
+          levels = c("STD", "MIDPT", "FAIR", "SL_MIDPT", "SL_FAIR", "SUGGEST"),
+          default = "SUGGEST",
+          tags = "train",
+          depends = quote(search == "kdtree")
         ),
         approx = p_dbl(default = 0, tags = "train"),
         eps_cl = p_dbl(0, tags = c("required", "train"))
       )
-      # add deps
-      param_set$add_dep("bucketSize", "search", CondEqual$new("kdtree"))
-      param_set$add_dep("splitRule", "search", CondEqual$new("kdtree"))
 
       super$initialize(
         id = "clust.optics",
         feature_types = c("logical", "integer", "numeric"),
         predict_types = "partition",
         param_set = param_set,
-        properties = c("partitional", "exclusive", "complete"),
+        properties = c("density", "exclusive", "complete"),
         packages = "dbscan",
         man = "mlr3cluster::mlr_learners_clust.optics",
         label = "OPTICS Clustering"
@@ -51,15 +51,15 @@ LearnerClustOPTICS = R6Class("LearnerClustOPTICS",
   private = list(
     .train = function(task) {
       pv = self$param_set$get_values(tags = "train")
-      m = invoke(dbscan::optics, x = task$data(), .args = remove_named(pv, "eps_cl"))
-      m = insert_named(m, list(data = task$data()))
+      data = task$data()
+      m = invoke(dbscan::optics, x = data, .args = remove_named(pv, "eps_cl"))
+      m = insert_named(m, list(data = data))
       m = invoke(dbscan::extractDBSCAN, object = m, eps_cl = pv$eps_cl)
 
       if (self$save_assignments) {
         self$assignments = m$cluster
       }
-
-      return(m)
+      m
     },
 
     .predict = function(task) {
@@ -69,5 +69,5 @@ LearnerClustOPTICS = R6Class("LearnerClustOPTICS",
   )
 )
 
-#' @include aaa.R
-learners[["clust.optics"]] = LearnerClustOPTICS
+#' @include zzz.R
+register_learner("clust.optics", LearnerClustOPTICS)