+++ R/PredictionDataClust.R
@@ -28,7 +28,6 @@ check_prediction_data.PredictionDataClust = function(pdata, ...) { # nolint
   pdata
 }
 
-
 #' @export
 is_missing_prediction_data.PredictionDataClust = function(pdata, ...) { # nolint
   miss = logical(length(pdata$row_ids))
@@ -89,4 +88,20 @@ filter_prediction_data.PredictionDataClust = function(pdata, row_ids, ...) {
   }
 
   pdata
+}
+
+#' @export
+create_empty_prediction_data.TaskClust = function(task, learner) {
+  predict_types = mlr_reflections$learner_predict_types[["clust"]][[learner$predict_type]]
+
+  pdata = list(
+    row_ids = integer(),
+    partition = integer()
+  )
+
+  if ("prob" %in% predict_types) {
+    pdata$prob = matrix(integer())
+  }
+
+  set_class(pdata, "PredictionDataClust")
 }