+++ R/as_prediction_clust.R
@@ -36,13 +36,11 @@ as_prediction_clust = function(x, ...) {
   UseMethod("as_prediction_clust")
 }
 
-
 #' @rdname as_prediction_clust
 #' @export
 as_prediction_clust.PredictionClust = function(x, ...) { # nolint
   x
 }
-
 
 #' @rdname as_prediction_clust
 #' @export
@@ -52,7 +50,7 @@ as_prediction_clust.data.frame = function(x, ...) { # nolint
 
   if (length(prob_cols)) {
     if (!all(startsWith(prob_cols, "prob."))) {
-      stopf("Table may only contain columns 'row_ids', 'partition' as well as columns prefixed with 'prob.' for class probabilities")
+      stopf("Table may only contain columns 'row_ids', 'partition' as well as columns prefixed with 'prob.' for class probabilities") # nolint
     }
     prob = as.matrix(x[, prob_cols, with = FALSE])
     colnames(prob) = substr(colnames(prob), 6L, nchar(colnames(prob)))