+++ tests/testthat/test_mlr_learners_clust_bico.R
@@ -1,7 +1,7 @@
-skip_if_not_installed("apcluster")
+skip_if_not_installed("stream")
 
 test_that("autotest", {
-  learner = mlr3::lrn("clust.ap", s = apcluster::negDistMat(r = 2L))
+  learner = mlr3::lrn("clust.bico")
   expect_learner(learner)
   result = run_autotest(learner)
   expect_true(result, info = result$error)
@@ -9,22 +9,22 @@ test_that("autotest", {
 
 test_that("Learner properties are respected", {
   task = tsk("usarrests")
-  learner = lrn("clust.ap")
+  learner = lrn("clust.bico")
   expect_learner(learner, task)
 
   # test on multiple paramsets
   parset_list = list(
-    list(s = apcluster::negDistMat(r = 2L)),
-    list(s = apcluster::linSimMat, details = TRUE, q = 0.5),
-    list(s = apcluster::expSimMat, lam = 0.5, nonoise = TRUE, includeSim = TRUE),
-    list(s = apcluster::corSimMat, convits = 50L, maxits = 500L)
+    list(k = 5),
+    list(k = 5, space = 5L),
+    list(k = 5, space = 5L, p = 5L),
+    list(k = 5, space = 5L, p = 5L, iterations = 5L)
   )
 
   for (i in seq_along(parset_list)) {
     parset = parset_list[[i]]
     learner$param_set$values = parset
 
-    suppressWarnings({p = learner$train(task)$predict(task)})
+    p = learner$train(task)$predict(task)
     expect_prediction_clust(p)
 
     if ("complete" %in% learner$properties) {