+++ tests/testthat/test_mlr_learners_clust_birch.R
@@ -1,7 +1,7 @@
-skip_if_not_installed("apcluster")
+skip_if_not_installed("stream")
 
 test_that("autotest", {
-  learner = mlr3::lrn("clust.ap", s = apcluster::negDistMat(r = 2L))
+  learner = mlr3::lrn("clust.birch", threshold = 0.1, branching = 8L, maxLeaf = 20L)
   expect_learner(learner)
   result = run_autotest(learner)
   expect_true(result, info = result$error)
@@ -9,22 +9,21 @@ test_that("autotest", {
 
 test_that("Learner properties are respected", {
   task = tsk("usarrests")
-  learner = lrn("clust.ap")
+  learner = lrn("clust.birch", threshold = 0.1, branching = 8L, maxLeaf = 20L)
   expect_learner(learner, task)
 
   # test on multiple paramsets
   parset_list = list(
-    list(s = apcluster::negDistMat(r = 2L)),
-    list(s = apcluster::linSimMat, details = TRUE, q = 0.5),
-    list(s = apcluster::expSimMat, lam = 0.5, nonoise = TRUE, includeSim = TRUE),
-    list(s = apcluster::corSimMat, convits = 50L, maxits = 500L)
+    list(threshold = 0.1, branching = 8L, maxLeaf = 20L),
+    list(threshold = 0.2, branching = 4L, maxLeaf = 10, maxMem = 2L),
+    list(threshold = 0.3, branching = 12L, maxLeaf = 5L, outlierThreshold = 0.3)
   )
 
   for (i in seq_along(parset_list)) {
     parset = parset_list[[i]]
     learner$param_set$values = parset
 
-    suppressWarnings({p = learner$train(task)$predict(task)})
+    p = learner$train(task)$predict(task)
     expect_prediction_clust(p)
 
     if ("complete" %in% learner$properties) {